Dragsters: a screenplay
An idea I have for a screenplay about a motorcycle gang who are also drag queens. It's clearly a comedy with some pretty simple plot lines encompassing great characters with distinct emotional development and an easy to understand story that nevertheless has enough twists and turns to keep you guessing.